autonomous-financial-analyst/requirements.txt
# ── Agent Framework ──────────────────────────────────────────────────────────
crewai==1.9.3

# ── Core Framework ──────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.1
pydantic-settings>=2.2.1

# ── LLM ──────────────────────────────────────────────────────────────────────
openai>=1.30.1

# ── Market Data ───────────────────────────────────────────────────────────────
yfinance>=0.2.40

# ── News ─────────────────────────────────────────────────────────────────────
newsapi-python==0.2.7

# ── Sentiment / NLP ──────────────────────────────────────────────────────────
transformers>=4.41.0
torch>=2.3.0
sentencepiece>=0.2.0

# ── Technical Analysis ───────────────────────────────────────────────────────
pandas-ta==0.4.71b0
ta>=0.11.0

# ── Vector Database ───────────────────────────────────────────────────────────
faiss-cpu>=1.8.0

# ── Data Processing ───────────────────────────────────────────────────────────
pandas>=2.2.0
numpy>=2.0.0
scipy>=1.13.0

# ── Charts ────────────────────────────────────────────────────────────────────
plotly>=5.22.0

# ── Frontend ─────────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.30
alembic>=1.13.0

# ── Scheduler ────────────────────────────────────────────────────────────────
apscheduler>=3.10.0

# ── HTTP / Utilities ─────────────────────────────────────────────────────────
httpx>=0.27.0
python-dotenv>=1.0.0
tenacity>=8.3.0
python-multipart>=0.0.9
slowapi>=0.1.9

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
